requirements.txt
Pillow
pytesseract
pytest
requests
boto3
